perf(hipblaslt): productionize gfx90c prompt GEMM catalogs Treat ordinary FP16 compute as native FP16 accumulation throughout hipBLASLt problem construction, cached-problem updates, input typing, and client references. Add tuned gfx90c prompt GEMM mappings for FP16 HHH and rocBLAS FP32 QK workloads, and pin their catalog selections with unit coverage.

+++ projects/hipblaslt/clients/common/include/testing_auxiliary.hpp
@@ -3329,7 +3329,7 @@ inline void testing_aux_tensile_host_func(const Arguments& arg)
                 == rocisa::DataType::Float);
     ASSERT_TRUE(rocComputeType_to_tensile_type(rocblaslt_compute_f32_fast_bf16)
                 == rocisa::DataType::Float);
-    ASSERT_TRUE(rocComputeType_to_tensile_type(rocblaslt_compute_f16) == rocisa::DataType::Float);
+    ASSERT_TRUE(rocComputeType_to_tensile_type(rocblaslt_compute_f16) == rocisa::DataType::Half);
     ASSERT_TRUE(rocComputeType_to_tensile_type(rocblaslt_compute_f16_pedantic)
                 == rocisa::DataType::Half);
     ASSERT_TRUE(rocComputeType_to_tensile_type(rocblaslt_compute_f32) == rocisa::DataType::Float);
